pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "caissAI"
authors = [
  {name = "Matthieu Daviaud", email = "matthieu.daviaud@gmail.com"},
]
version = "0.1.0"
description = "A python project to annotate and comment chess games"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
dependencies = [
    "beartype>=0.22.9",
    "maia2>=0.9",
    "memoization>=0.4.0",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "psutil>=7.2.2",
    "pyarrow>=23.0.1",
    "python-chess>=1.999",
    "pyyaml>=6.0.3",
    "pyzstd>=0.19.1",
    "torch>=2.10.0",
]

[project.scripts]
caissAI = "src.caissAI.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest-cov>=7.0.0",
]
docs = [
    "sphinx>=9.1.0",
    "sphinx-rtd-theme>=3.1.0",
]
lint = [
    "mypy>=1.19.1",
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py313"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests", "src/caissAI"]
addopts = [
    "--doctest-modules",
    "--cov=src/caissAI",
    "--cov-report=term-missing",
    "--cov-report=html",
]
pythonpath = ["."]

[tool.coverage.run]
source = ["src/caissAI"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
    "raise NotImplementedError",
]
skip_empty = true

[tool.mypy]
python_version = "3.13"
strict = true

[[tool.mypy.overrides]]
module = ["maia2", "maia2.*", "memoization"]
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src/caissAI"]

[tool.hatch.build.targets.wheel.force-include]
"src/caissAI/py.typed" = "caissAI/py.typed"

[dependency-groups]
dev = [
    "pandas-stubs>=3.0.0.260204",
    "types-psutil>=7.2.2.20260130",
    "types-tqdm>=4.67.3.20260303",
]
doc = [
    "sphinx>=9.1.0",
]
lint = [
    "mypy>=1.19.1",
    "ruff>=0.15.5",
]
